requirements.txt
certifi==2018.10.15
chardet==3.0.4
codecov==2.0.15
coverage==4.5.1
diff-match-patch==20121119
Django==2.1.2
django-excel==0.0.10
django-import-export==1.1.0
et-xmlfile==1.0.1
flake8==3.6.0
idna==2.7
jdcal==1.4
Jinja2==2.10
lml==0.0.4
lxml==4.2.5
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
odfpy==1.3.6
openpyxl==2.5.9
pkg-resources==0.0.0
pycodestyle==2.4.0
pyecharts-jupyter-installer==0.0.3
pyexcel==0.5.9.1
pyexcel-ezodf==0.3.4
pyexcel-handsontable==0.0.2
pyexcel-io==0.5.9.1
pyexcel-ods3==0.5.2
pyexcel-pygal==0.0.2
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.6
pyflakes==2.0.0
pygal==2.4.0
pytz==2018.6
PyYAML==3.13
requests==2.20.0
tablib==0.12.1
texttable==1.4.0
unicodecsv==0.14.1
urllib3==1.24
xlrd==1.1.0
xlwt==1.3.0
